Elevation HeightMap introduced
Instead of saving elevation data in the context of a single job, elevation data is saved in a key/value store, the HeightMap class, so that it can be loaded and re-used for any jobs covering the same area. Cuts down massively on calls to endpoint.  HeightMap uses lat,long coordinates as string keys, but does cut down to a maximum decimal precision (currently 4) so we're not saving supersampled points.
Doesn't work so great with ProcessPoolExecutor, yet.